Table complete, but missing clickable row  functionality
