+++ src/pages/manager/schedule/index.tsx
@@ -1,5 +1,4 @@
 import Layout from "../../../components/Layout/Layout";
-import ShiftSchedule from "../../../components/ManagerComponents/shiftSchedule/ShiftSchedule";
 import UnassignedShiftList from "../../../components/ManagerComponents/UnassignedShiftList/unassignedShiftList";
 
 export default function SchedulePage() {
